Combine test_deploy and test_upgrade_charms
